refactor: ProfitResult, PlayerProfitResult double -> int로 타입 변경

+++ src/main/java/blackjack/dto/ProfitResult.java
@@ -2,6 +2,6 @@ package blackjack.dto;
 
 import java.util.List;
 
-public record ProfitResult(double dealerProfit, List<PlayerProfitResult> playerProfits) {
+public record ProfitResult(int dealerProfit, List<PlayerProfitResult> playerProfits) {
 }
